+++ app/api/Files.scala
@@ -1910,7 +1910,7 @@ class Files @Inject()(
       }
       case None => {
         Logger.error("Error getting file " + id)
-        InternalServerError
+        NotFound
       }
     }
   }
@@ -1921,11 +1921,14 @@ class Files @Inject()(
       case Some(file) => {
         files.setStatus(id, FileStatus.PROCESSED)
         sinkService.logFileUnarchiveEvent(file, user)
+
+        // Only increment download date, to avoid immediate auto-archive
+        files.incrementDownloads(id, user, true)
         Ok(toJson(Map("status" -> "success")))
       }
       case None => {
         Logger.error("Error getting file " + id)
-        InternalServerError
+        NotFound
       }
     }
   }
@@ -1941,7 +1944,7 @@ class Files @Inject()(
       }
       case None => {
         Logger.error("Error getting file " + id)
-        InternalServerError
+        NotFound
       }
     }
   }
@@ -1957,7 +1960,7 @@ class Files @Inject()(
       }
       case None => {
         Logger.error("Error getting file " + id)
-        InternalServerError
+        NotFound
       }
     }
   }