+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 - Added folder and folder id to api datasets files list [#34](https://github.com/clowder-framework/clowder/issues/34)
 
+### Added
+- Ability to queue archive / unarchive for full datasets.
+
 ### Fixed
 - RabbitMQ will not use connection if None exists
 - previews returns 404 if preview is not found